:goal_net: handle timeout and validation errors in ToFhirRejectionHandler
